MindSPONGE/applications/research/UFold/requirements.txt
mindspore-gpu==1.8.1
munch==2.5.0
scipy==1.7.3
numpy==1.21.6
pandas==1.3.5
scikit_learn==1.0.2
